chore(*): automatically handle line endings
Adding a `.gitattributes` file to automatically handle end of line
conversions. This will help make things more consistent and help
prevent issues where something works on one machine but not another.

An example of this is the maven wrapper which fails if you use the
bash shell script and the `maven-wrapper.properties` file has CRLF
line endings.  This change will ensure that the file always has LF
line endings.